DAOtests can be written now DataFormat final class containing constants for data parsing

+++ pong/src/main/java/entity/data/DataSet.java
@@ -7,6 +7,6 @@ public interface DataSet {
 
     int getNumberOfGames();
     int getNumberOfWins();
-    int getPlayTime();
+    long getPlayTime();
     // TODO: more dataset functionality later
 }
